reorganize methods in modules
