+++ pysages/ssages/methods/funn.py
@@ -1,0 +1,76 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (c) 2021: SSAGES Team (see LICENSE.md)
+
+from collections import namedtuple
+
+import jax.numpy as np
+from jax.numpy import linalg
+
+from pysages.ssages.grids import get_index
+from pysages.ssages.methods import NNSamplingMethod, generalize
+from pysages.nn.models import mlp
+from pysages.nn.objectives import PartialRBObjective
+from pysages.nn.optimizers import LevenbergMaquardtBayes
+from pysages.nn.training import trainer
+
+# ======== #
+#   FUNN   #
+# ======== #
+
+
+class FUNNState(
+    namedtuple(
+        "FUNNState",
+        ("bias", "nn", "hist", "Fsum", "F", "Wp", "Wp_"),
+    )
+):
+    def __repr__(self):
+        return repr("PySAGES " + type(self).__name__)
+
+
+class FUNN(NNSamplingMethod):
+    def __call__(self, snapshot, helpers):
+        N = np.asarray(self.kwargs.get('N', 200))
+        return _funn(snapshot, self.cv, self.grid, self.topology, N, helpers)
+
+
+def _funn(snapshot, cv, grid, topology, N, helpers):
+    dt = snapshot.dt
+    dims = grid.shape.size
+    natoms = np.size(snapshot.positions, 0)
+    indices, momenta = helpers
+    model = mlp(grid.shape, dims, topology)
+    train = trainer(model, PartialRBObjective(), LevenbergMaquardtBayes(), np.zeros(dims))
+
+    def initialize():
+        bias = np.zeros((natoms, dims))
+        hist = np.zeros(grid.shape, dtype=np.uint32)
+        Fsum = np.zeros((*grid.shape, dims))
+        F = np.zeros(dims)
+        Wp = np.zeros(dims)
+        Wp_ = np.zeros(dims)
+        return FUNNState(bias, model.parameters, hist, Fsum, F, Wp, Wp_)
+
+    def update(state, rs, vms, ids):
+        # Compute the collective variable and its jacobian
+        ξ, Jξ = cv(rs, indices(ids))
+        #
+        θ = train(state.nn, state.Fsum / state.hist).θ
+        Q = model.apply(θ, ξ)
+        #
+        p = momenta(vms)
+        Wp = linalg.tensorsolve(Jξ @ Jξ.T, Jξ @ p)
+        dWp_dt = (1.5 * Wp - 2.0 * state.Wp + 0.5 * state.Wp_) / dt
+        #
+        I_ξ = get_index(grid, ξ)
+        N_ξ = state.hist[I_ξ] + 1
+        F_ξ = state.Fsum[I_ξ] + dWp_dt + Q
+        hist = state.hist.at[I_ξ].set(N_ξ)
+        Fsum = state.Fsum.at[I_ξ].set(F_ξ)
+        F = F_ξ / np.maximum(N_ξ, N)
+        #
+        bias = np.reshape(-Jξ.T @ F, state.bias.shape)
+        #
+        return FUNNState(bias, θ, hist, Fsum, F, Wp, state.Wp)
+
+    return snapshot, initialize, generalize(update)
